Add VirtualSize, enforce block vsize, and tighten types
Introduce a VirtualSize trait and implement it for Block, Transaction,
Input, OutputId and Output. Block::verify now checks the virtual size
and rejects blocks over 1_000_000 bytes with InvalidBlockSize.

Replace variable-length signatures with a fixed Signature = [u8;64]
alias and change OutputId.index to u8. Update miner, ledger, vm and
tests to use the new types and remove some unused code/prints.

+++ src/core/ledger/in_mem.rs
@@ -47,7 +47,7 @@ impl InMemoryLedger {
             let tx_id = tx.hash::<Blake2s256>();
             for (i, output) in tx.outputs.iter().enumerate() {
                 self.utxo_set
-                    .insert(OutputId::new(tx_id, i), output.clone());
+                    .insert(OutputId::new(tx_id, i as u8), output.clone());
             }
         }
         Ok(())
